+++ examples/change_listener/main.go
@@ -31,7 +31,7 @@ type Listener struct {
 }
 
 // Listen is called on entity changes.
-func (l *Listener) Listen(evt *ecs.EntityEvent) {
+func (l *Listener) Listen(evt ecs.EntityEvent) {
 	// Just prints out what the event is about.
 	// This could also be a method of a type that manages events.
 	// Could use e.g. filters to distribute events to interested/registered systems.