+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 This change was necessary to get the same performance as before, despite the more heavyweight implementation of the now 256 bits `Mask`.
 * Component and resource IDs are now opaque types instead of type aliases for `uint8` (#329)
 * Restructures `EntityEvent` to remove redundant information and better handle relation changes (#333)
+* World event listener function signature changed from `func(*EntityEvent)` to `func(EntityEvent)` (#333)
 
 ### Features